Adjust Npgsql dep version to match Wyn image
Wyn bundles Npgsql 6.0.7.0 as of this writing.  Even adding our own
copy of Npgsql, the Wyn version sometimes loads first.  (There may or
may not be a load time race condition there.)  If the Wyn version loads
first, then our DLL will not load as it depends on a later version.